src/pam_python.c: wrap Py C calls in PyGILState_Ensure/Release()
avoids segfaults when the extension is loaded from a PAM module.

We have to ensure that PyGILState_Ensure is called only after
Py_Initialize() is called or that leads to another segfault.

Py_Initialize() is currently called within a call to get_pamHandle. So
the calls in get_pamHandle() AFTER the initialization are wrapped and
protected.

cleanup_pamHandle() also needs to be protected as it seems to tun
asynchronously in some other thread.

All the other functions which make PyC calls are called through
call_handler(), so it is probably sufficient to wrap the portion of
call_handler() AFTER a pamhandle is obtained.

https://bugs.python.org/issue44913